requirements.txt
hyperliquid-python-sdk>=0.9.0
python-dotenv>=1.0.0
